requirements.txt
numpy
pillow
scikit-learn
scikit-image